+++ src/app/code/community/Dhl/Versenden/Helper/Address.php
@@ -38,10 +38,10 @@ class Dhl_Versenden_Helper_Address extends Mage_Core_Helper_Abstract
 
         if (preg_match($this->getStreetSplitter(), $street, $matches)) {
             // Pattern A
-            if (isset($matches[self::OPTION_A_STREET_NAME]) && !empty($matches[self::OPTION_A_STREET_NAME])) {
+            if (isset($matches[self::OPTION_A_STREET_NAME]) && $matches[self::OPTION_A_STREET_NAME] !== '') {
                 $result['street_name'] = trim($matches[self::OPTION_A_STREET_NAME]);
 
-                if (isset($matches[self::OPTION_A_HOUSE_NUMBER]) && !empty($matches[self::OPTION_A_HOUSE_NUMBER])) {
+                if (isset($matches[self::OPTION_A_HOUSE_NUMBER]) && $matches[self::OPTION_A_HOUSE_NUMBER] !== '') {
                     $result['street_number'] = trim($matches[self::OPTION_A_HOUSE_NUMBER]);
                 }
 
@@ -52,10 +52,10 @@ class Dhl_Versenden_Helper_Address extends Mage_Core_Helper_Abstract
                 }
 
                 // Pattern B
-            } elseif (isset($matches[self::OPTION_B_STREET_NAME]) && !empty($matches[self::OPTION_B_STREET_NAME])) {
+            } elseif (isset($matches[self::OPTION_B_STREET_NAME]) && $matches[self::OPTION_B_STREET_NAME] !== '') {
                 $result['street_name'] = trim($matches[self::OPTION_B_STREET_NAME]);
 
-                if (isset($matches[self::OPTION_B_HOUSE_NUMBER]) && !empty($matches[self::OPTION_B_HOUSE_NUMBER])) {
+                if (isset($matches[self::OPTION_B_HOUSE_NUMBER]) && $matches[self::OPTION_B_HOUSE_NUMBER] !== '') {
                     $result['street_number'] = trim($matches[self::OPTION_B_HOUSE_NUMBER]);
                 }